Move pipeline routes under ee src tree
