+++ forge/ee/routes/index.js
@@ -11,6 +11,8 @@ module.exports = async function (app) {
     }
     await app.register(require('./sharedLibrary'), { logLevel: app.config.logging.http })
 
+    await app.register(require('./pipeline'), { prefix: '/api/v1', logLevel: app.config.logging.http })
+
     // Important: keep SSO last to avoid its error handling polluting other routes.
     await app.register(require('./sso'), { logLevel: app.config.logging.http })
 }
