+++ forge/routes/api/application.js
@@ -192,79 +192,4 @@ module.exports = async function (app) {
             reply.code(404).send({ code: 'not_found', error: 'Not Found' })
         }
     })
-
-    /**
-     * List all pipelines within an Application
-     * @name /api/v1/application/:id/pipelines
-     * @memberof forge.routes.api.application
-     */
-    app.get('/:applicationId/pipelines', {
-        // TODO: What permissions are required here?
-        preHandler: app.needsPermission('team:projects:list')
-    }, async (request, reply) => {
-        const pipelines = await app.db.models.Pipeline.byApplicationId(request.application.hashid)
-        if (pipelines) {
-            reply.send({
-                count: pipelines.length,
-                pipelines: await app.db.views.Pipeline.pipelineList(pipelines)
-            })
-        } else {
-            reply.code(404).send({ code: 'not_found', error: 'Not Found' })
-        }
-    })
-
-    /**
-     * Create a new Pipeline within an Application
-     * @name /api/v1/application/:id/pipelines
-     * @memberof forge.routes.api.application
-     */
-    app.post('/:applicationId/pipelines', {
-        // TODO: What permissions are required here?
-        preHandler: app.needsPermission('team:projects:list')
-    }, async (request, reply) => {
-        const team = await request.teamMembership.getTeam()
-        const name = request.body.name?.trim()
-
-        let pipeline
-        try {
-            pipeline = await app.db.models.Pipeline.create({
-                name,
-                ApplicationId: request.application.id
-            })
-        } catch (err) {
-            console.error(err)
-            return reply.status(500).send({ code: 'unexpected_error', error: err.toString() })
-        }
-
-        await app.auditLog.Team.application.pipeline.created(request.session.User, null, team, request.application, pipeline)
-
-        reply.send(app.db.views.Pipeline.pipeline(pipeline))
-    })
-
-    /**
-     * Delete a Pipeline from an Application
-     * @name /api/v1/application/:id/pipelines
-     * @memberof forge.routes.api.application
-     */
-    app.delete('/:applicationId/pipelines/:pipelineId', {
-        // TODO: What permissions are required here?
-        preHandler: app.needsPermission('team:projects:list')
-    }, async (request, reply) => {
-        const team = await request.teamMembership.getTeam()
-        const pipelineId = request.params.pipelineId
-        const pipeline = await app.db.models.Pipeline.byId(pipelineId)
-        const stages = await pipeline.stages()
-
-        if (stages.length > 0) {
-            // delete stages too
-            for (let i = 0; i < stages.length; i++) {
-                stages[i].destroy()
-            }
-        }
-
-        await pipeline.destroy()
-        await app.auditLog.Team.application.pipeline.deleted(request.session.User, null, team, request.application, pipeline)
-
-        reply.send({ status: 'okay' })
-    })
 }
