+++ forge/routes/api/index.js
@@ -10,7 +10,6 @@ const Users = require('./users.js')
 const Team = require('./team.js')
 const TeamType = require('./teamType.js')
 const Application = require('./application.js')
-const Pipeline = require('./pipeline.js')
 const Project = require('./project.js')
 const Admin = require('./admin.js')
 const Settings = require('./settings.js')
@@ -36,7 +35,6 @@ module.exports = async function (app) {
     app.register(Team, { prefix: '/teams' })
     app.register(TeamType, { prefix: '/team-types' })
     app.register(Application, { prefix: '/applications' })
-    app.register(Pipeline, { prefix: '/pipelines' })
     app.register(Project, { prefix: '/projects' })
     app.register(Stack, { prefix: '/stacks' })
     app.register(Template, { prefix: '/templates' })
